downLoadReferenceImageError
