+++ src/utils/utils.ts
@@ -107,20 +107,6 @@ export const settings2GraphAIConfig = (
   return deepClean(config) ?? {};
 };
 
-export const getExtention = (contentType: string | null, url: string) => {
-  if (contentType?.includes("jpeg") || contentType?.includes("jpg")) {
-    return "jpg";
-  } else if (contentType?.includes("png")) {
-    return "png";
-  }
-  // Fall back to URL extension
-  const urlExtension = url.split(".").pop()?.toLowerCase();
-  if (urlExtension && ["jpg", "jpeg", "png"].includes(urlExtension)) {
-    return urlExtension === "jpeg" ? "jpg" : urlExtension;
-  }
-  return "png"; // default
-};
-
 // deepClean
 
 type Primitive = string | number | boolean | symbol | bigint;
